requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
langchain_community==0.3.9
langchain_core==0.3.21
openai==1.55.3
pydantic==2.10.2
pytest==8.3.4
python-dotenv==1.0.1
WTForms==3.2.1
